requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dj-database-url==3.1.0
Django
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.20.2
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-ollama==1.0.1
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.0
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.2
psycopg2==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyMuPDF==1.26.7
python-dotenv==1.2.1
PyYAML==6.0.3
qdrant-client==1.16.2
rich==14.3.1
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.12.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
whitenoise
beautifulsoup4